refactor: split diplomacy rows and production breakdown dialog (#3878)
Decompose diplomacy_panel_rows.dart into standing-chip, builder, and
power-comparison parts, and extract ProductionBreakdownTableBody from the
commodity breakdown dialog so each panel submodule stays under size targets
without behavior changes.

Refs #3878
