+++ app/lib/features/game/widgets/diplomacy_panel_rows.dart
@@ -6,86 +6,9 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-int? _outgoingSubsidyPercent(Game game, String payerId, String targetId) {
-  for (final s in game.subsidyStates) {
-    if (s.payerId == payerId && s.targetId == targetId) {
-      return s.percent;
-    }
-  }
-  return null;
-}
-
-void _appendDiscoveredFactionId(
-  Game game,
-  String humanPlayerId,
-  String id,
-  List<String> gpIds,
-  List<String> minorIds,
-  List<String> tribeIds,
-) {
-  if (id == humanPlayerId) {
-    return;
-  }
-  if (game.players.any((p) => p.id == id)) {
-    gpIds.add(id);
-    return;
-  }
-  if (game.minorNations.any((m) => m.id == id)) {
-    minorIds.add(id);
-    return;
-  }
-  if (game.tribes.any((t) => t.id == id)) {
-    tribeIds.add(id);
-  }
-}
-
-/// Display names of Great Powers the human GP boycotts through a colony Tribe
-/// it holds (Refs #3753 R12). Extracted so the standing-chip builder stays
-/// under the repo control-flow nesting-depth gate.
-List<String> _boycottVsNames(Game game, String humanPlayerId) {
-  final names = <String>[];
-  for (final b in game.boycottStates) {
-    if (b.gpId == humanPlayerId) {
-      names.add(game.playerById(b.targetGpId)?.displayName ?? b.targetGpId);
-    }
-  }
-  return names;
-}
-
-/// Display names of Great Powers boycotting the human GP from trading with a
-/// colony Tribe held by [colonyOfGpId] (Refs #3753 R12). Extracted so the
-/// standing-chip builder stays under the repo control-flow nesting-depth gate.
-List<String> _boycottedByNames(
-  Game game,
-  String humanPlayerId,
-  String colonyOfGpId,
-) {
-  final names = <String>[];
-  for (final b in game.boycottStates) {
-    if (b.gpId == colonyOfGpId && b.targetGpId == humanPlayerId) {
-      names.add(game.playerById(colonyOfGpId)?.displayName ?? colonyOfGpId);
-    }
-  }
-  return names;
-}
-
-({int? grant, int? subsidy}) _pendingEconomicAmounts(
-  List<DiplomaticOrder> list,
-  String targetId,
-) {
-  int? grant;
-  int? subsidy;
-  for (final o in list) {
-    if (o.targetFactionId != targetId) continue;
-    if (o.type == DiplomaticOrderType.grantAid) {
-      grant = o.amount;
-    }
-    if (o.type == DiplomaticOrderType.setSubsidy) {
-      subsidy = o.amount;
-    }
-  }
-  return (grant: grant, subsidy: subsidy);
-}
+part 'diplomacy_panel_rows_standing_chips.dart';
+part 'diplomacy_panel_rows_builder.dart';
+part 'diplomacy_panel_rows_power.dart';
 
 /// Faction type for display. SPEC/game/factions.md.
 enum FactionKind { greatPower, minor, tribe }
@@ -151,75 +74,6 @@ class DiplomaticStandingChips {
       overseasTileCount <= 0;
 
   bool get isNotEmpty => !isEmpty;
-}
-
-/// Derives the [DiplomaticStandingChips] for a faction row from current
-/// [game] state. SPEC/ui/diplomacy-panel.md § Diplomatic standing chip
-/// cluster (Refs #3753 R12). [purchasedTiles] is built once per render pass
-/// by the caller (e.g. [buildDiplomacyRows]) so the per-tile attribution scan
-/// is not repeated per row.
-DiplomaticStandingChips diplomaticStandingChips({
-  required Game game,
-  required String humanPlayerId,
-  required String factionId,
-  required FactionKind kind,
-  required DiplomacyRelation? relation,
-  required OvertureState? overture,
-  required PurchasedTileIndex purchasedTiles,
-}) {
-  final OvertureStage stage = overture?.stage ?? OvertureStage.none;
-  final treaty = <String>[];
-  if (stage.index >= OvertureStage.tradeConsulate.index) {
-    treaty.add(kDiplomacyChipConsulate);
-  }
-  if (stage.index >= OvertureStage.embassy.index) {
-    treaty.add(kDiplomacyChipEmbassy);
-  }
-  if (stage.index >= OvertureStage.nap.index) {
-    treaty.add(kDiplomacyChipNap);
-  }
-
-  String? colonyOfGpId;
-  if (kind == FactionKind.tribe) {
-    for (final c in game.colonyStates) {
-      if (c.tribeId == factionId) {
-        colonyOfGpId = c.colonyOfGpId;
-        break;
-      }
-    }
-  }
-  final bool isColonyOfHuman = colonyOfGpId == humanPlayerId;
-  if (isColonyOfHuman) {
-    treaty.add(kDiplomacyChipColony);
-  } else if (kind == FactionKind.minor && stage == OvertureStage.joinEmpire) {
-    treaty.add(kDiplomacyChipJoinEmpire);
-  }
-
-  List<String> boycottVs = const <String>[];
-  List<String> boycottedBy = const <String>[];
-  if (colonyOfGpId != null && isColonyOfHuman) {
-    boycottVs = _boycottVsNames(game, humanPlayerId);
-  } else if (colonyOfGpId != null) {
-    boycottedBy = _boycottedByNames(game, humanPlayerId, colonyOfGpId);
-  }
-
-  int overseasCount = 0;
-  if (kind != FactionKind.greatPower) {
-    for (final a in purchasedTiles.attributions) {
-      if (a.owningGpId == humanPlayerId && a.sourceFactionId == factionId) {
-        overseasCount++;
-      }
-    }
-  }
-  final int overseasShare = relation == null ? 0 : relation.score.round();
-
-  return DiplomaticStandingChips(
-    treatyLabels: treaty,
-    boycottVsNames: boycottVs,
-    boycottedByNames: boycottedBy,
-    overseasTileCount: overseasCount,
-    overseasSharePercent: overseasShare,
-  );
 }
 
 /// Diplomacy panel bottom mode-bar filter selection, per
@@ -296,301 +150,4 @@ class DiplomacyRowData {
   /// faction. SPEC/ui/diplomacy-panel.md § Diplomatic standing chip cluster
   /// (Refs #3753 R12).
   final DiplomaticStandingChips standingChips;
-}
-
-/// Default neutral first-contact standing surfaced for a discovered faction
-/// that has no persisted [DiplomacyRelation] yet. SPEC/ui/diplomacy-panel.md
-/// § Discovered factions → First-contact standing: `AT_PEACE`, score `50`,
-/// level `Neutral` (the same default used for game-start Minor relations).
-/// The `DiplomacyRelation` constructor already defaults to these values; the
-/// turn fields are pinned to the current turn so history-derived UI stays
-/// deterministic.
-DiplomacyRelation _defaultFirstContactRelation(
-  String humanPlayerId,
-  String factionId,
-  int currentTurn,
-) => DiplomacyRelation(
-  factionId1: humanPlayerId,
-  factionId2: factionId,
-  sinceTurn: currentTurn,
-  lastInteractionTurn: currentTurn,
-);
-
-/// Builds list of discovered factions and their available actions.
-///
-/// Discovery follows `knownDiplomaticTargetFactionIds`
-/// (SPEC/ui/diplomacy-panel.md § Discovered factions) for Great Powers and
-/// Minor Nations: an existing [DiplomacyRelation] or non-`unknown` tile
-/// visibility into a faction-owned province. A discovered faction without a
-/// persisted relation is surfaced with the default neutral first-contact
-/// standing.
-///
-/// **Tribes require first contact (Refs #3620):** a Tribe is surfaced only
-/// after first contact — a persisted GP↔Tribe relation **or** non-`unknown`
-/// tile visibility into a province that Tribe owns
-/// ([discoveredTribeIdsForFirstContact]). Sea-reachable colonial intel alone
-/// no longer surfaces a Tribe row (SPEC/ui/diplomacy-panel.md § Tribes require
-/// first contact).
-List<DiplomacyRowData> buildDiplomacyRows(
-  Game game,
-  MapTopology topology,
-  String humanPlayerId,
-  Orders currentOrders,
-) {
-  final view = buildPlayerView(game, topology, humanPlayerId);
-  // Built once per render pass so the per-tile attribution scan that backs the
-  // overseas-holdings standing chip (Refs #3753 R12 / R8) is not repeated per
-  // row. SPEC/ui/diplomacy-panel.md § Diplomatic standing chip cluster.
-  final purchasedTiles = PurchasedTileIndex.fromGame(game);
-  final discoveredIds = <String>{
-    ...view.diplomacyByOtherId.keys,
-    ...knownDiplomaticTargetFactionIds(
-      view: view,
-      game: game,
-      topology: topology,
-    ),
-  };
-  // First-contact gate for Tribes (Refs #3620): drop tribe ids that are only
-  // sea-reachable colonial intel (present in `knownDiplomaticTargetFactionIds`
-  // without a persisted relation or tile visibility). A tribe is contacted iff
-  // a GP↔Tribe relation is persisted or the GP holds non-`unknown` tile
-  // visibility into a province it owns.
-  final contactedTribeIds = discoveredTribeIdsForFirstContact(
-    view: view,
-    game: game,
-  );
-  discoveredIds.removeWhere((id) {
-    final isTribe = game.tribes.any((t) => t.id == id);
-    if (!isTribe) return false;
-    final hasRelation = view.diplomacyByOtherId.containsKey(id);
-    return !hasRelation && !contactedTribeIds.contains(id);
-  });
-  final currentTurn = game.worldState.turnState.turnNumber;
-  final factionMembership = DiplomacyFactionMembership.from(game);
-  final actionsByTarget = <String, List<DiplomaticPanelAction>>{};
-  for (final id in discoveredIds) {
-    if (id == humanPlayerId) continue;
-    actionsByTarget[id] = enumerateDiplomaticPanelActionsForTarget(
-      game: game,
-      topology: topology,
-      playerId: humanPlayerId,
-      targetId: id,
-      currentOrders: currentOrders,
-      factionMembership: factionMembership,
-    );
-  }
-
-  final gpIds = <String>[];
-  final minorIds = <String>[];
-  final tribeIds = <String>[];
-  for (final id in discoveredIds) {
-    _appendDiscoveredFactionId(
-      game,
-      humanPlayerId,
-      id,
-      gpIds,
-      minorIds,
-      tribeIds,
-    );
-  }
-
-  // GPs: sort by military power desc, then province count desc. SPEC/ui/diplomacy-panel.md.
-  gpIds.sort((a, b) {
-    final strA = aggregateMilitaryStrengthForPlayer(game, a);
-    final strB = aggregateMilitaryStrengthForPlayer(game, b);
-    final cmp = strB.compareTo(strA);
-    if (cmp != 0) return cmp;
-    final provA = provinceCountOwnedBy(game, a);
-    final provB = provinceCountOwnedBy(game, b);
-    return provB.compareTo(provA);
-  });
-
-  final pendingByTarget = <String, Set<DiplomaticOrderType>>{};
-  final pendingOvertureStageByTarget = <String, OvertureStage>{};
-  final pendingList =
-      currentOrders.diplomaticOrdersByPlayerId[humanPlayerId] ?? [];
-  for (final o in pendingList) {
-    pendingByTarget.putIfAbsent(o.targetFactionId, () => {}).add(o.type);
-    if (o.type == DiplomaticOrderType.establishOverture &&
-        o.overtureStage != null) {
-      pendingOvertureStageByTarget[o.targetFactionId] = o.overtureStage!;
-    }
-  }
-
-  String displayNameFor(String id) {
-    final p = game.playerById(id);
-    if (p != null) return p.displayName;
-    for (final m in game.minorNations) {
-      if (m.id == id) return m.displayName ?? id;
-    }
-    for (final t in game.tribes) {
-      if (t.id == id) return t.displayName ?? id;
-    }
-    return id;
-  }
-
-  List<DiplomacyRowData> rows = [];
-  final playerPower = greatPowerPowerScore(game, humanPlayerId);
-  for (final id in gpIds) {
-    final econ = _pendingEconomicAmounts(pendingList, id);
-    rows.add(
-      DiplomacyRowData(
-        factionId: id,
-        displayName: displayNameFor(id),
-        kind: FactionKind.greatPower,
-        relation:
-            view.diplomacyByOtherId[id] ??
-            _defaultFirstContactRelation(humanPlayerId, id, currentTurn),
-        overture: getOverture(game, humanPlayerId, id),
-        actions: actionsByTarget[id] ?? const <DiplomaticPanelAction>[],
-        powerScore: greatPowerPowerScore(game, id),
-        playerPowerScore: playerPower,
-        pendingOrderTypes: pendingByTarget[id] ?? {},
-        pendingOvertureStage: pendingOvertureStageByTarget[id],
-        activeSubsidyPercent: _outgoingSubsidyPercent(game, humanPlayerId, id),
-        pendingGrantAmount: econ.grant,
-        pendingSubsidyPercent: econ.subsidy,
-        standingChips: diplomaticStandingChips(
-          game: game,
-          humanPlayerId: humanPlayerId,
-          factionId: id,
-          kind: FactionKind.greatPower,
-          relation:
-              view.diplomacyByOtherId[id] ??
-              _defaultFirstContactRelation(humanPlayerId, id, currentTurn),
-          overture: getOverture(game, humanPlayerId, id),
-          purchasedTiles: purchasedTiles,
-        ),
-      ),
-    );
-  }
-  minorIds.sort((a, b) => (displayNameFor(a)).compareTo(displayNameFor(b)));
-  for (final id in minorIds) {
-    final econ = _pendingEconomicAmounts(pendingList, id);
-    rows.add(
-      DiplomacyRowData(
-        factionId: id,
-        displayName: displayNameFor(id),
-        kind: FactionKind.minor,
-        relation:
-            view.diplomacyByOtherId[id] ??
-            _defaultFirstContactRelation(humanPlayerId, id, currentTurn),
-        overture: getOverture(game, humanPlayerId, id),
-        actions: actionsByTarget[id] ?? const <DiplomaticPanelAction>[],
-        powerScore: null,
-        playerPowerScore: null,
-        pendingOrderTypes: pendingByTarget[id] ?? {},
-        pendingOvertureStage: pendingOvertureStageByTarget[id],
-        activeSubsidyPercent: _outgoingSubsidyPercent(game, humanPlayerId, id),
-        pendingGrantAmount: econ.grant,
-        pendingSubsidyPercent: econ.subsidy,
-        standingChips: diplomaticStandingChips(
-          game: game,
-          humanPlayerId: humanPlayerId,
-          factionId: id,
-          kind: FactionKind.minor,
-          relation:
-              view.diplomacyByOtherId[id] ??
-              _defaultFirstContactRelation(humanPlayerId, id, currentTurn),
-          overture: getOverture(game, humanPlayerId, id),
-          purchasedTiles: purchasedTiles,
-        ),
-      ),
-    );
-  }
-  tribeIds.sort((a, b) => (displayNameFor(a)).compareTo(displayNameFor(b)));
-  for (final id in tribeIds) {
-    final econ = _pendingEconomicAmounts(pendingList, id);
-    rows.add(
-      DiplomacyRowData(
-        factionId: id,
-        displayName: displayNameFor(id),
-        kind: FactionKind.tribe,
-        relation:
-            view.diplomacyByOtherId[id] ??
-            _defaultFirstContactRelation(humanPlayerId, id, currentTurn),
-        overture: getOverture(game, humanPlayerId, id),
-        actions: actionsByTarget[id] ?? const <DiplomaticPanelAction>[],
-        powerScore: null,
-        playerPowerScore: null,
-        pendingOrderTypes: pendingByTarget[id] ?? {},
-        pendingOvertureStage: pendingOvertureStageByTarget[id],
-        activeSubsidyPercent: _outgoingSubsidyPercent(game, humanPlayerId, id),
-        pendingGrantAmount: econ.grant,
-        pendingSubsidyPercent: econ.subsidy,
-        standingChips: diplomaticStandingChips(
-          game: game,
-          humanPlayerId: humanPlayerId,
-          factionId: id,
-          kind: FactionKind.tribe,
-          relation:
-              view.diplomacyByOtherId[id] ??
-              _defaultFirstContactRelation(humanPlayerId, id, currentTurn),
-          overture: getOverture(game, humanPlayerId, id),
-          purchasedTiles: purchasedTiles,
-        ),
-      ),
-    );
-  }
-  return rows;
-}
-
-/// Computes the relative Great Power power-comparison percentage used by
-/// `SPEC/ui/diplomacy-panel.md` § Power comparison percentage.
-///
-/// Formula: `round(((gpPowerScore - playerPowerScore) / max(playerPowerScore,
-/// 1)) * 100)`. The `max(.., 1)` guard prevents division-by-zero when the
-/// human player's score is `0`, yielding a finite percentage.
-///
-/// Returns the integer percentage (positive when the GP is stronger, negative
-/// when weaker, zero when equal).
-int powerComparisonPercent(int gpPowerScore, int playerPowerScore) {
-  final int denom = math.max(playerPowerScore, 1);
-  final double ratio = (gpPowerScore - playerPowerScore) / denom;
-  return (ratio * 100).round();
-}
-
-/// Formats a `powerComparisonPercent` integer for display per
-/// `SPEC/ui/diplomacy-panel.md` § Power comparison percentage:
-/// `+N%` when positive, `−N%` (U+2212 minus) when negative, `0%` when zero.
-///
-/// The minus sign is the unicode `MINUS SIGN` (U+2212) to match the mockup,
-/// not the ASCII hyphen-minus (U+002D).
-String formatPowerComparisonPercent(int pct) {
-  if (pct > 0) return '+$pct%';
-  if (pct < 0) return '\u2212${-pct}%';
-  return '0%';
-}
-
-/// Display-only strength tier derived from a `powerComparisonPercent` value
-/// per `SPEC/ui/diplomacy-panel.md` § Relative power line. The tier is a
-/// UI-only label and never feeds AI war-desire or any logic-package model.
-enum PowerComparisonTier {
-  vastlyInferior,
-  inferior,
-  roughlyEqual,
-  superior,
-  vastlySuperior,
-}
-
-/// Maps a `powerComparisonPercent` integer to its [PowerComparisonTier] per
-/// the boundary table in `SPEC/ui/diplomacy-panel.md` § Relative power line:
-///
-/// | `pct` range | Tier |
-/// |-------------|------|
-/// | `pct >= +31` | vastlySuperior |
-/// | `+11 .. +30` | superior |
-/// | `−10 .. +10` | roughlyEqual |
-/// | `−30 .. −11` | inferior |
-/// | `pct <= −31` | vastlyInferior |
-///
-/// Boundaries are inclusive on the side shown (e.g. `+10` is roughlyEqual,
-/// `+11` is superior). Extreme values (e.g. `+4900` when the player score is
-/// near zero) clamp into [PowerComparisonTier.vastlySuperior] without a cap.
-PowerComparisonTier powerComparisonTier(int pct) {
-  if (pct <= -31) return PowerComparisonTier.vastlyInferior;
-  if (pct <= -11) return PowerComparisonTier.inferior;
-  if (pct <= 10) return PowerComparisonTier.roughlyEqual;
-  if (pct <= 30) return PowerComparisonTier.superior;
-  return PowerComparisonTier.vastlySuperior;
 }
