+++ app/lib/features/game/widgets/production_commodity_breakdown_dialog.dart
@@ -16,6 +16,8 @@ import '../../../widgets/ct_resource_cell.dart';
 import '../../../widgets/ct_spacing.dart';
 import '../../../widgets/resource_icon.dart';
 
+part 'production_commodity_breakdown_dialog_table.dart';
+
 /// Minimum viewport width at which the breakdown dialog drops the
 /// trailing horizontal `Scrollbar` + `SingleChildScrollView` and lets the
 /// 7-column `DataTable` lay out without horizontal scroll.
@@ -117,7 +119,7 @@ class ProductionCommodityBreakdownDialog extends ConsumerStatefulWidget {
   ConsumerState<ProductionCommodityBreakdownDialog> createState() =>
       _ProductionCommodityBreakdownDialogState();
 
-  static String _phaseColumnLabel(
+  static String phaseColumnLabel(
     AppLocalizations l10n,
     EconomyPreviewStockpilePhase phase,
   ) {
@@ -135,7 +137,7 @@ class ProductionCommodityBreakdownDialog extends ConsumerStatefulWidget {
     };
   }
 
-  static String _formatDelta(int v) {
+  static String formatDelta(int v) {
     if (v > 0) return '+$v';
     return '$v';
   }
@@ -155,70 +157,6 @@ class _ProductionCommodityBreakdownDialogState
   void dispose() {
     _horizontalScrollController.dispose();
     super.dispose();
-  }
-
-  TextStyle _headingStyle(BuildContext context) {
-    final TextStyle base =
-        Theme.of(context).textTheme.labelSmall ?? const TextStyle(fontSize: 11);
-    return base.copyWith(
-      color: EditorialMonoclePalette.muted,
-      fontWeight: FontWeight.w600,
-    );
-  }
-
-  TextStyle _sectionHeaderStyle(BuildContext context) {
-    final TextStyle base =
-        Theme.of(context).textTheme.labelSmall ?? const TextStyle(fontSize: 11);
-    return base.copyWith(
-      color: EditorialMonoclePalette.muted,
-      fontWeight: FontWeight.w600,
-      fontFeatures: const <FontFeature>[FontFeature.enable('smcp')],
-    );
-  }
-
-  TextStyle _commodityNameStyle(BuildContext context) {
-    final TextStyle base =
-        Theme.of(context).textTheme.bodySmall ?? const TextStyle(fontSize: 12);
-    return base.copyWith(color: EditorialMonoclePalette.fg);
-  }
-
-  TextStyle _deltaCellStyle(BuildContext context, int value) {
-    final TextStyle base =
-        Theme.of(context).textTheme.labelMedium ?? const TextStyle(fontSize: 12);
-    final Color? color = CtResourceCell.deltaColor(value);
-    return base.copyWith(
-      color: color ?? EditorialMonoclePalette.muted,
-      fontFamilyFallback: const <String>['SF Mono', 'Menlo', 'monospace'],
-      fontFeatures: const <FontFeature>[FontFeature.tabularFigures()],
-    );
-  }
-
-  Widget _deltaCell(BuildContext context, int value) {
-    return Text(
-      ProductionCommodityBreakdownDialog._formatDelta(value),
-      maxLines: 1,
-      style: _deltaCellStyle(context, value),
-    );
-  }
-
-  Widget _sectionHeaderCell(BuildContext context, String label) {
-    return DecoratedBox(
-      decoration: BoxDecoration(
-        border: Border(
-          bottom: BorderSide(
-            color: EditorialMonoclePalette.accentDim,
-            width: 1,
-          ),
-        ),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.only(bottom: CtSpacing.xs),
-        child: Text(
-          label.toUpperCase(),
-          style: _sectionHeaderStyle(context),
-        ),
-      ),
-    );
   }
 
   @override
@@ -273,213 +211,12 @@ class _ProductionCommodityBreakdownDialogState
       ),
     ];
 
-    int phaseValue(String commodityId, EconomyPreviewStockpilePhase phase) {
-      return phaseDeltas[phase]?[commodityId] ?? 0;
-    }
-
-    int rowTotal(String commodityId) {
-      var t = 0;
-      for (final p in EconomyPreviewStockpilePhase.values) {
-        t += phaseValue(commodityId, p);
-      }
-      return t;
-    }
-
-    final phaseColCount = EconomyPreviewStockpilePhase.values.length;
-    final dividerColor =
-        EditorialMonoclePalette.accentDim.withValues(alpha: 0.5);
-
-    // Tightened column padding shared by both viewport paths. Declared here
-    // so the wide-path full-width distribution math can subtract the exact
-    // chrome (`horizontalMargin` on both outer edges + `columnSpacing`
-    // between columns) that the DataTable reinstates around/between columns.
-    // SPEC/ui/production-commodity-breakdown-dialog.md § Layout. Refs #2862
-    // S8c / C6 + #3509.
-    const double tableColumnSpacing = 24;
-    const double tableHorizontalMargin = 12;
-
-    // Builds the 7-column breakdown table. When [columnContentWidths] is
-    // non-null (wide path), every header label and cell child is pinned to an
-    // explicit per-column width so the table fills the full dialog content
-    // column; when null (narrow path) children keep intrinsic sizing inside
-    // the horizontal scroll viewport.
-    // SPEC/ui/production-commodity-breakdown-dialog.md § Layout. Refs #3509.
-    DataTable buildDataTable(List<double>? columnContentWidths) {
-      Widget sizedCell(int columnIndex, Widget child) {
-        if (columnContentWidths == null) return child;
-        return SizedBox(width: columnContentWidths[columnIndex], child: child);
-      }
-
-      // Header wrappers carry stable keys so layout tests can read each
-      // column's distributed width directly. Refs #3509.
-      Widget sizedHeader(int columnIndex, Widget child) {
-        if (columnContentWidths == null) return child;
-        return SizedBox(
-          key: ValueKey<String>('prodBreakdownHeaderCol_$columnIndex'),
-          width: columnContentWidths[columnIndex],
-          child: child,
-        );
-      }
-
-      var commodityRowIndex = 0;
-
-      List<DataRow> rowsFor(List<Commodity> list) {
-        return list.map((c) {
-          final total = rowTotal(c.id);
-          final name = c.displayName ?? c.id;
-          final rowShade = commodityRowIndex.isEven
-              ? Colors.transparent
-              : EditorialMonoclePalette.surface.withValues(alpha: 0.4);
-          commodityRowIndex += 1;
-          return DataRow(
-            color: WidgetStatePropertyAll<Color?>(rowShade),
-            cells: [
-              DataCell(
-                sizedCell(
-                  0,
-                  Row(
-                    children: [
-                      ResourceIcon(commodityId: c.id, size: 16),
-                      const SizedBox(width: 4),
-                      Expanded(
-                        child: Text(
-                          name,
-                          maxLines: 1,
-                          overflow: TextOverflow.ellipsis,
-                          style: _commodityNameStyle(context),
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-              ...EconomyPreviewStockpilePhase.values.indexed.map(
-                (entry) => DataCell(
-                  sizedCell(
-                    1 + entry.$1,
-                    _deltaCell(context, phaseValue(c.id, entry.$2)),
-                  ),
-                ),
-              ),
-              DataCell(
-                sizedCell(phaseColCount + 1, _deltaCell(context, total)),
-              ),
-            ],
-          );
-        }).toList();
-      }
-
-      return DataTable(
-        headingRowHeight: 40,
-        dataRowMinHeight: 32,
-        dataRowMaxHeight: 48,
-        dividerThickness: 1,
-        // Tighten DataTable's default column padding so the 7-column
-        // breakdown fits inside the wide-viewport `CtDialogShell` body
-        // without requiring a horizontal scroll affordance at viewport
-        // width >= 900 dp. The narrow path still wraps the table in a
-        // `Scrollbar` + horizontal `SingleChildScrollView` so the same
-        // table remains reachable below the threshold (see the conditional
-        // below). SPEC/ui/production-commodity-breakdown-dialog.md § Layout.
-        // Refs #2862 S8c / C6 + #3509.
-        columnSpacing: tableColumnSpacing,
-        horizontalMargin: tableHorizontalMargin,
-        headingRowColor: WidgetStatePropertyAll<Color?>(
-          EditorialMonoclePalette.surfaceLite,
-        ),
-        border: TableBorder(
-          horizontalInside: BorderSide(color: dividerColor),
-        ),
-        headingTextStyle: _headingStyle(context),
-        columns: [
-          DataColumn(
-            label: sizedHeader(0, Text(l10n.production_breakdown_commodity)),
-          ),
-          ...EconomyPreviewStockpilePhase.values.indexed.map(
-            (entry) => DataColumn(
-              label: sizedHeader(
-                1 + entry.$1,
-                Text(
-                  ProductionCommodityBreakdownDialog._phaseColumnLabel(
-                    l10n,
-                    entry.$2,
-                  ),
-                  softWrap: true,
-                ),
-              ),
-            ),
-          ),
-          DataColumn(
-            label: sizedHeader(
-              phaseColCount + 1,
-              Text(l10n.production_breakdown_total),
-            ),
-          ),
-        ],
-        rows: [
-          for (final (label, commodities) in sections)
-            if (commodities.isNotEmpty) ...[
-              DataRow(
-                cells: [
-                  DataCell(sizedCell(0, _sectionHeaderCell(context, label))),
-                  ...List<DataCell>.generate(
-                    phaseColCount + 1,
-                    (i) => DataCell(
-                      sizedCell(1 + i, const SizedBox.shrink()),
-                    ),
-                  ),
-                ],
-              ),
-              ...rowsFor(commodities),
-            ],
-        ],
-      );
-    }
-
     final viewportWidth = MediaQuery.of(context).size.width;
     final isWideViewport =
         viewportWidth >= kProductionBreakdownDialogWideViewportThreshold;
     final resolvedMaxWidth = isWideViewport
         ? kProductionBreakdownDialogWideMaxWidth
         : kProductionBreakdownDialogNarrowMaxWidth;
-
-    final Widget tableBody;
-    if (isWideViewport) {
-      // Wide path: distribute the full dialog content-column width across the
-      // 7 columns (Commodity weighted larger; phase columns + Total share the
-      // remainder evenly) so the table fills the dialog with no trailing gap
-      // (owner decision B on #3509). No Scrollbar / horizontal scroll
-      // affordance is mounted — the table is sized to fit the content column
-      // exactly. SPEC/ui/production-commodity-breakdown-dialog.md § Layout.
-      // Refs #3509.
-      tableBody = LayoutBuilder(
-        builder: (BuildContext context, BoxConstraints constraints) {
-          final widths = productionBreakdownWideColumnContentWidths(
-            availableWidth: constraints.maxWidth,
-            phaseColumnCount: phaseColCount,
-            columnSpacing: tableColumnSpacing,
-            horizontalMargin: tableHorizontalMargin,
-          );
-          return buildDataTable(widths);
-        },
-      );
-    } else {
-      // Narrow path unchanged: the intrinsic-width table lives inside a
-      // visible horizontal `Scrollbar` + `SingleChildScrollView` so every
-      // column stays reachable on small viewports.
-      // SPEC/ui/production-commodity-breakdown-dialog.md § Layout. Refs #2862
-      // S8c / C6.
-      final Widget scrollableTable = SingleChildScrollView(
-        controller: _horizontalScrollController,
-        scrollDirection: Axis.horizontal,
-        child: buildDataTable(null),
-      );
-      tableBody = Scrollbar(
-        controller: _horizontalScrollController,
-        thumbVisibility: true,
-        child: scrollableTable,
-      );
-    }
 
     return CtDialogShell(
       maxWidth: resolvedMaxWidth,
@@ -495,7 +232,13 @@ class _ProductionCommodityBreakdownDialogState
             ),
           ),
           CtGap.m,
-          tableBody,
+          ProductionBreakdownTableBody(
+            l10n: l10n,
+            sections: sections,
+            phaseDeltas: phaseDeltas,
+            isWideViewport: isWideViewport,
+            horizontalScrollController: _horizontalScrollController,
+          ),
           CtGap.ml,
           Align(
             alignment: Alignment.centerRight,
